removed VmInterface
